+++ lib/oli/delivery/attempts/client_evaluation.ex
@@ -1,0 +1,7 @@
+defmodule Oli.Delivery.Attempts.ClientEvaluation do
+
+  @enforce_keys [:input, :score, :out_of, :feedback]
+  @derive Jason.Encoder
+  defstruct [:input, :score, :out_of, :feedback]
+
+end